requirements.txt
alembic==1.4.3
bcrypt==3.2.0
boto3==1.16.18
botocore==1.19.18
certifi==2020.12.5
cffi==1.14.3
chardet==4.0.0
click==7.1.2
dnspython==2.0.0
email-validator==1.1.2
Faker==4.14.2
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-JWT-Extended==3.25.0
Flask-Login==0.5.0
flask-marshmallow==0.13.0
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
Mako==1.1.3
MarkupSafe==1.1.1
marshmallow==3.8.0
marshmallow-sqlalchemy==0.23.1
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==1.7.1
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
requests==2.25.1
s3transfer==0.3.3
six==1.15.0
SQLAlchemy==1.3.19
text-unidecode==1.3
urllib3==1.26.2
Werkzeug==1.0.1
WTForms==2.3.3
WTFormValidation==1.0.0
